Pipeline example
